tasks/vision/classification/resnet/requirements.txt
tensorflow==2.12.0
numpy==1.23.5
tqdm==4.66.1
Pillow==10.0.1
tensorflow_datasets==4.9.2